pyproject.toml
[tool.poetry]
name = "stvimhelper"
version = "0.1.0"
description = "Vim helper scripts for SmartThings tools"
authors = ["Nick Stevens <nick@bitcurry.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
atlassian-python-api = "^3.14.0"
click = "^8.0.1"
PyGithub = "^1.55"

[tool.poetry.dev-dependencies]
mypy = "^0.910"

[tool.poetry.scripts]
stvimhelper= "stvimhelper:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
